+++ buildSrc/build.gradle.kts
@@ -1,0 +1,8 @@
+import org.gradle.kotlin.dsl.`kotlin-dsl`
+
+plugins {
+    `kotlin-dsl`
+}
+repositories {
+    jcenter()
+}
