+++ settings.gradle.kts
@@ -1,0 +1,8 @@
+include(
+    ":app",
+    ":visionai-core",
+    ":visionai-shared",
+    ":tflite:shared",
+    ":tflite:styletransfer",
+    ":tflite:bodysegmentation"
+)
