+++ tflite/bodysegmentation/proguard-rules.pro
