+++ tflite/bodysegmentation/src/main/AndroidManifest.xml
