+++ tflite/bodysegmentation/src/main/assets/segmentation_gpu.tflite
