+++ tflite/bodysegmentation/src/main/res/values/strings.xml
