+++ tflite/shared/build.gradle
