+++ tflite/shared/src/main/AndroidManifest.xml
