+++ tflite/shared/src/main/res/values/strings.xml
