+++ tflite/styletransfer/src/main/AndroidManifest.xml
