+++ tflite/styletransfer/src/main/assets/detect.tflite
