+++ tflite/styletransfer/src/main/assets/flowers.jpg
