+++ tflite/styletransfer/src/main/assets/happy_mood_spring.jpg
