+++ tflite/styletransfer/src/main/assets/labelmap.txt
