+++ tflite/styletransfer/src/main/assets/lion.jpg
