+++ tflite/styletransfer/src/main/assets/picasso_self.jpg
