+++ tflite/styletransfer/src/main/assets/starry_night.jpg
