+++ tflite/styletransfer/src/main/assets/style_predict_quantized_256.tflite
