+++ tflite/styletransfer/src/main/assets/style_transfer_quantized_dynamic.tflite
