+++ tflite/styletransfer/src/main/java/com/bendenen/visionai/tflite/styletransfer/step/ArtisticStyleTransferStepResult.kt
