+++ tflite/styletransfer/src/main/res/values/strings.xml
